requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
rich>=13.7.0
flask>=3.0.0
